requirements.txt
numpy==1.24.2
pandas==1.5.3
psycopg2_binary==2.9.10
python-dotenv==1.0.1
python_bcrypt==0.3.2
streamlit==1.40.2